Crate dim_core

Dim is a media manager written in rust. It uses Diesel as the ORM and rocket for the http/s server

The project is split up into several crates:

database - Holds all the database models including some frequently used db operations

routes - All of the routes that we expose over http are stored in there

scanners - The filesystem scanner and daemon code is located here ffmpeg that is used by several parts of dim

Building

Dim can easily be built with cargo build –release. When built with –release, build.rs will compile the web ui and embed it into dim.

To run

Dim can be ran using docker, by pulling vgarleanu/dim-server, or locally. If ran locally, make sure PostgreSQL is running with the password for postgres: dimpostgres

Testing

To test run make test in the root, or cargo test in the root of each module including the root dir.

Re-exports

pub use routes::settings::get_global_settings;

pub use routes::settings::init_global_settings;

pub use routes::settings::set_global_settings;

pub use routes::settings::GlobalSettings;

Modules

core

Module contains our core initialization logic.

errors

Module contains all the error definitions used in dim, and returned by the web-service.

fetcher

Contains the code for fetching assets like posters and stills.

inspect

Inspect api for Result type

reactor

Sqlite CDC implementation

routes

Contains all of the routes exposed by the webapi.

scanner

New generation scanner infrastructure. Module contains all the code for the new generation media scanner.

stream_tracking

Contains the fairing which tracks streams across rest api

streaming

Contains all the logic needed for streaming and on-the-fly transcoding.

tree

Tree-like structure for representing directories of files.

utils

Various utilities

Functions

setup_logging

Function builds a logger drain that drains to a json file located in logs/ and also to stdout.