dim_core

Module scanner

New generation scanner infrastructure. Module contains all the code for the new generation media scanner.

Re-exports

pub use error::Error;

Modules

daemon

error

movie

tv_show

Structs

WorkUnit

Traits

MediaMatcher

Trait that must be implemented by a media matcher. Matchers are responsible for fetching their own external metadata but it is provided a metadata provider at initialization time.

Functions

format_path

Function formats the path where assets are stored.

get_subfiles

Function recursively walks the paths passed and returns all files in those directories. FIXME: THIS IS NOT ASYNC-SAFE!!! NOTE: I’ve noticed that walking a directory mounted over ssh is very slow, 80 files in like 300 seconds. Doubt theres a way to fix this but we could alliviate the UX-degradation by sending the files over a channel instead of returning them at once.

insert_mediafiles

parse_filenames

start

start_custom